Fix manual runs that are hitting an assert 0x162 and make default local run to be time  (#15242)
## Description

An assert 0x162 is showing up when running the benchmark tests against
the local service. The problem is that instead of loading an existing
container, the test was creating a new one. Also, we want only the
regular time test to run during the individual test debugging
experience. Running both would trigger the provider to be reset after
each test and would cause errors during the second test.

Alternatives considered: 
1) Use provider = getTestObjectProvider({ resetAfterEach: false });   in
each test.
2) Change describeE2EDocs to only reset after all instead of after each