[GC stress] Fix error codes on process.exit (#15234)
Apparently, the exit error code in node cannot be negative. Calling
process.exit(-1) sets the code to 255, proces.exit(-2) sets it to 254
and so on, i.e., it wraps around. That's the reason the failures in the
GC stress tests weren't repoted correctly.
Updated the tests to use 255 for regular test failures and 254 for GC
failures.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test