[GC stress] Added another none activity to increase the chances of incremental summary (#15256)
## Description
One of the interesting scenarios for GC is incremental summary / GC for
data stores and DDSes. Bascially, if a data store or DDS did not change
since the last summary, it will re-use the GC / summary data from the
previous summary. This goes down a different code path and has known to
cause GC bugs.
This change increases the chances of that by adding another activity
which doesn't do anything, thereby increasing the chances of doing an
incremental summary for a data store.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test