Fix undo of overlapping remove of multiple segments (#15258)
## Description

Fixes an issue spotted in merge-tree while doing the recent undo/redo
asymptotics improvements. There's more context
[here](https://github.com/microsoft/FluidFramework/pull/14737#discussion_r1160281800)
or in the test case. The gist of the problem is that we didn't preserve
the order of local references when sliding them, which undo-redo cares
about.

Resolves
[AB#4031](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4031).

This also addresses some of the cleanup feedback on
https://github.com/microsoft/FluidFramework/pull/14737, as it was the
simplest way to fix this issue.