refactor(client-debugger): Remove "devtoolsDisposed" event from IFluidDevtools (#15266)
This event was added assuming that the visualization tooling would
require direct access to the Devtools object, and would need to know
when the consuming app disposed of the instance. Since the only consumer
of the Devtools object will be the consuming application, there is no
need for this event.