Refactor Rebaser (#15267)
## Description

This PR pulls the `rebaseBranch` and `rebaseChange` functions out of the
`Rebaser` class and makes them free functions in rebaser/utils.ts.
`Rebaser` is deleted; it was not a very useful class. It only had one
member (which is now passed into the free functions) and it was awkward
to instantiate in many places as well as being potentially confusing
alongside other "rebaser" classes (like `ChangeRebaser`). The
documentation of `rebaseBranch` has been improved as well, and there's
some minor cleanup in the implementation.