Fix cached snapshot id (#15275)
## Description

1.) Snapshot id in afr could be different from id of first tree in
snapshot. So use correct id as version instead of first snapshot.
For gitrest the id of snapshot is different from id of first tree in
snapshot. So we need to cache with proper id and not with id of first
tre. Although with the previous changes, it will still hit the cache
because we use the same id but it was not the correct one.
In my previous PR, I changed it to use the id of first tree which was
not correct way.
2.) Add scenario name so that we can know when getWholeFlatSummary event
is recorded, on snapshot load or later.