Refactor LocalIntervalCollection into indexes (#15246)
## Description

This refactors the implementation of LocalIntervalCollection to closer
align with the planned public API--see
[design](https://github.com/microsoft/FluidFramework/pull/15091/files).

I also removed the functionality of `addConflictResolver`, which was
marked deprecated in #15081. Those APIs have resulted in no-ops since FF
version 0.40 which is below LTS, so this change should not impact
customers.

This otherwise doesn't touch the public API, though we will eventually
export some of this functionality somehow--tasks under
[AB#3251](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3251)
track that.