Various small refactors in SharedTreeCore (#15336)
* The undo/redo type parameter in `trackCommit`, `applyChange`,
`submitCommit`, etc. is now required. This is because a developer adding
a new call to any of those functions needs to think about how their new
case interacts with undo/redo in order to avoid introducing bugs. The
functions that had a "skip" parameter have had that parameter collapsed
into the undo/redo type parameter for brevity. An explicit value of
`undefined` now means "don't track WRT undo/redo".
* `addLocalChange` is no longer responsible for either tracking
undo/redo commits or emitting change events. It never did so when called
from `applyChange`, but rather only when called from `applyStashedOp` -
and the return signature of `addLocalChange` makes it very easy for
`applyStashedOp` to do those two things itself. This reduces the
parameters passed to `addLocalChange` and makes its scope of
responsibility clearer.
* `applyChange` is now private rather than protected. Perhaps it will be
re-exposed in the future, but there's no need for it be exposed
currently.
* `applyChange` now requires a revision tag parameter. This forces
callers to double check whether they already have a revision handy, or
want to make a new one (which is just a call to `mintRevisionTag()`).
* Added and corrected documentation for the relevant functions