[GC stress] Renamed data stores and updated Readme (#15340)
## Description
Renamed the following data stores to better reflect what they do:
- DataStoreNonCollab -> SingleCollabDataStore - Only one client
collaborates on these data stores.
- DataStoreCollab -> MultiCollabDataStore - Multiple clients collborate
on these data stores.
- 
Also updated the Readme to reflect new names + attachment blobs.


[AB#2139](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2139)

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.