Update typebox (#15272)
## Description

Updates typebox to latest. There are a few nice features of the newer
version, but the main one that's relevant for our persisted types is
[composition of intersect and
union](https://github.com/sinclairzx81/typebox/blob/master/changelog/0.26.0.md#intersect-and-union-now-compose).

It also makes a few breaking changes to schema construction for a few
types. There are some details
[here](https://github.com/sinclairzx81/typebox/issues/373), but the gist
of it is we should avoid constraining types until they're fully
composed. So this refactors a few schemas defined in SchemaIndexFormat
to follow that recommendation.