Fix GC tests failing with default read mode (#15354)
As part of
[#3904](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3904),
container is loaded in "read" mode be default. Some GC tests were
failing with this change so a workaround was added to load the container
in "write" mode for tests tests.

This PR removes that workaround and makes the tests work with "read"
mode containers. There are 2 ways these tests are fixed:
1. The test in `gcInactiveNodes.spec.ts` requires the container to be in
"write" mode so that the GC data in base snapshot is loaded and marks
unreferenced nodes as inactive. So, updated the test to send an op and
wait for "write" connection.
2. The other tests were not testing the inactive / sweep ready state but
rather tombstone / deleted state. For these, the container does not need
to load in "write" mode. So, for these tests removed the expected
inactive / sweep ready errors since they won't happen in loaded
containers anymore.


[AB#3982](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3982)