Make more descriptive SharedTreeBranch events (#15369)
## Description

This PR adds more information to the events emitted by SharedTreeBranch.
In an upcoming refactor, the root local branch in
SharedTreeCore/EditManager will be replaced with a SharedTreeBranch, and
these events will allow SharedTreeCore to respond more intelligently to
changes to the branch, regardless of the source of those changes (rather
than all changes to the branch being routed through SharedTreeCore in
the first place, as they currently are).

Some of the information emitted by the branch events was already being
computed by `rebaseBranch`, so `rebaseBranch` now returns this extra
information rather than it being recomputed outside. `rebaseBranch` also
early-outs in a new case: when the branch being rebased has the same
(revision tag as) commits as the target branch it is rebasing over. This
is a nice optimization, and also in the future this will allow the root
local branch to rebase over the trunk without emitting any change events
(otherwise, it would emit a change event, and while that change event
would be "empty", we have consumers of the edit manager that rely on no
event being emitted at all).

I also took this opportunity to add a unit test suite for
`rebaseBranch`.