IdCompressor in the runtime (#14340)
## Description

This PR exists as a point of discussion for implementation details
around integrating the IdCompressor into the runtime. This has been
forked from #13398.

## Details

### What to Review
The IdCompressor has already been used in production in legacy
SharedTree and is thoroughly tested. Anything in
`packages/runtime/container-runtime/src/id-compressor` can be ignored
for the purpose of this review as it's just being moved to the runtime
folder.

### Overview
The idea is to have a container runtime level IdCompressor that is
shared across DDSs. DDS authors will be able to generate compressed Ids
using the runtime-level compressor. Upon DDS op submission, the runtime
checks the compressor's state to see if any compressed ops were
generated that need to be sent. If there are `IdAllocation` ops, the
runtime puts an `IdAllocation` op in front of the DDS's op.

### Stashed Ops
The idea is to "resume" the IdCompressor that the stashed client was
using by including the current compressor state in its contents. This is
done using `localOpMetadata` until the container is closed using
`closeAndGetPendingLocalState`. At this point, the `localOpMetadata` is
copied into `contents.stashedState` so that the stashed state is stored
for rehydrating the compressor. This state is deleted before sending the
op over the wire for other compressors to synchronize.

### Enablement and Compat
This feature is turned off by default and must be explicitly enabled.
The feature gate is `enableRuntimeIdCompressor`. The idea here is to do
the same thing we did with compression where we make breaking changes
without them turning on until reaching proper saturation.

### Exposure to DDS
The `IdCompressor` exists inside the container runtime but needs to be
accessed by DDSs. I've added accessors to the IdCompressor from the
container runtime down to the DataStoreRuntime level.

I've introduced two new interfaces to `runtime-definitions`:
`IdCompressorCore` and `IdCompressor`. `IdCompressorCore` contains
methods that shouldn't be exposed to a DDS author, but should be exposed
at the container runtime level to manage the `IdCompressor`'s state.
Upon retrieving the `IdCompressor`, the DDS has access to all methods
present in the `IdCompressor` interface.

### Finalizing Id Ranges
`IdRange`s are included inside `DataStoreOp`'s contents at the
`SharedObjectCore` level inside of `submitLocalMessage`. It must be
included in the op's contents so that ops that are resubmitted through
`replayPendingStates` correctly resubmit the `IdRange`. If appended at a
higher level, the `IdRange` would be lost on replay.

### Summarization
The `IdCompressor`'s state is added to the summary inside
`ContainerRuntime::addContainerStateToSummary`. I've added it as a blob
(should it be a tree?). The blob is fetched as part of the
`ContainerRuntime`'s constructor.