feat: Add check changeset command (#15320)
## Description

Adds a `flub check changesets` command which we will use in CI to
determine whether a changeset was added. The existing changesets tools
don't seem to work for us, hence this command.

I expanded the Repository class that contains our git logic to support
checking for changed files, directories, packages, and release groups.
This is not all used in this PR, but will be used in the future.