[GC stress] Don't fail run for non GC failures (#15367)
## Description
Currently, the stress test runs report as failed for non-GC failures
where the test exists due to uncaught exception.
This PR changes that so that the run will report failure only because of
GC failures. Added an `exitCode` property to test results where a runner
exits with non-zero exitCode so it can be debugged if needed.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.