test(client-debugger-view): Add a SharedCell example to the test app (#15345)
I'm hoping to expand the set of DDSs we use in our test app to let us
better test the data visualization features. This PR adds a SharedCell
to the test app.

The cell powers an "Emoji Button" that just toggles between different
smileys on click. I'm definitely open to alternative suggestions if
folks have them :)


![image](https://user-images.githubusercontent.com/54606601/235010278-5cb44d6d-a9d7-4354-acac-ef67198613db.png)