feat: Add check buildVersion command (#15392)
We recently had an issue (see
https://github.com/microsoft/FluidFramework/pull/15381#issuecomment-1530727486)
where new packages were added in main and then when merged to next, they
had versions that did not match the release group/workspace version.

The `flub check buildVersion` command can be used in CI to verify that
all packages in a release group have the correct version. We will run
this command in the release group policy tasks in CI. With this in
place, we will at least have some protection against this situation.

Note that this does not check dependency versions - only the version in
package.json. Dependency versions are handled by syncpack and the
upcoming workspace protocol work.

This is not implemented as a repo policy handler because those run on
the whole repo and would require more config to implement.