[GC stress] Better failure handling and log unexpected errors (#15286)
## Description
Fixed instances where failures were not propagated all the way up to the
runner. Improved the error handlers and event handlers. Any unexpected
errors are now logged.


[AB#2137](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2137)

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.