[GC stress] Adjust the various timeouts to account for snapshot cache expiry (#15410)
## Description
https://github.com/microsoft/FluidFramework/pull/14454 added snapshot
cache expiry to the GC stress tests. However, the inactive timeout and
sweep timeout was not adjusted accordingly. This means there is a small
window where a client may be able to load inactive / sweep ready
objects. We see 3 instances of `sweepReadyObject_Loaded` since #14454
was merged.

This change adjusts the time based on snapshot cache expiry. It does the
following for odsp runs only:
- Reduced snapshot cache expiry to 5 mins.
- Reduced faultInjection max time to 10 mins so that snapshot cache
expiry is half of it. The idea is that on average half the sessions that
fail because of faultInjection will load from cached snapshot and other
half will not.
- Increased inactive timeout to 25 mins and sweep timeout to 26 mins.
This means there is a 5 mins buffer time (same as before) between
session expiry (15 mins) + snapshot cache expiry (5 mins) to inactive
timeout (25 mins). This buffer should be good enough for ops to
propagate to all clients.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.