feat(bump): Support workspace protocol in release and bump tools (#15053)
## Summary

This PR encompasses three major changes:

1. Adds a config setting to control the default interdependency ranges
for release groups.
2. Refactors the `bump`-related functions into `set` functions.
3. Adds support for the workspace protocol.

## Default interdependency range setting

Adds a configuration setting to release groups that controls their
interdependency range type. This is a companion change to #15158 and
similar changes. That change (and related changes) allowed setting
explicit range types in the CI pipeline and used that capability to
release builds with `~` interdependencies.

On the other hand, this change adds a setting to the fluid-build
configuration that is then used as a default in relevant places rather
than the `^` default that was there previously. With this change, the
`flub release` and related commands will use the configured
interdependency range type unless one is passed as an argument.

I introduced a couple of new types to help consolidate disparate string
union types. I also took this opportunity to deprecate and replace the
poorly named `exactDepType` flag. It's been replaced by the
`interdependencyRange` flag, a name which we use throughout the code to
refer to this concept.

## Refactoring bump and version handling

The existing code mixed _determining the version to set_ with _setting
the version_. This led to unwieldy 'bump' functions that have been
refactored or in some cases removed. The primary replacement functions
are `setVersion` and `setPackageDependencies`.

`setVersion` can be used to set a package or release group to a
particular version. It properly reads the default config setting
described above and has parameters that can be used to override the
defaults. Once it determines the correct interdependency ranges, it
calls `setPackageDependencies` to set the interdependencies across the
release group.

The `bumpReleaseGroup` function was refactored to call `setVersion`. It
may be removed in a future PR.

## Workspace protocol support

The [workspace
protocol](https://pnpm.io/workspaces#workspace-protocol-workspace)
enables pnpm to always link the workspace version of a package. It also
enables pnpm to help enforce that versions are consistent between
interdependent packages.

Because of our custom version schemes, we use custom bump and version
tools. I've updated these tools to partially understand the workspace
protocol. The `bump` command's `interdependencyRange` flag can be set to
values like `workspace:^`, enabling us to set workspace ranges in CI and
elsewhere.

Fixes
[AB#3422](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3422).

### Examples

These examples can be run on the FluidFramework repo (with the changes
in this PR applied, of course).

Sets the version of the build-tools release group to 0.16.0, and use
"workspace:*" for dependencies on other packages within the workspace.

```shell
flub bump build-tools --exact 0.16.0 -d "workspace:*"
```

Bumps the client release group to the next minor version using the
default interdependency range configured for the client release group in
the fluidBuild config file.

```shell
flub bump client -t minor
```

## Other changes

- Removed the unused packageVersion files from all the build-tools
packages. They were unused.

## Breaking changes

BREAKING CHANGE: The `--exactDepType` flag in the `bump` command no
longer has a default value. It has also been deprecated. It has been
replaced by the `--interdependencyType` flag. The deprecated flag will
be removed in an upcoming release.