ci: Create draft release notes on tag push (#15288)
This change adds a new GitHub actions pipeline that will automate the
release creation on tag push, including creating the report files and
uploading them to the release. The OCE/release engineer will just need
to hit "publish" on the release in GitHub to make it public. For patch
releases we'll even generate the release notes automatically. For minors
and majors that's still a manual process.

## How it works

When a release tag is pushed to the repo, the workflow is triggered. It
first installs the Fluid build-tools, then uses the `flub release
fromTag` command to load some release metadata into an environment
variable. Once loaded, it checks out the tagged commit and runs `flub
release report` to generate release reports. It also uses auto-changelog
to create a changelog for patch releases (only patches). All the
artifacts are uploaded for debugging purposes.

Once the artifacts are created, the workflow creates a GitHub release
and attaches the release reports to it. The release is a draft so that
the release engineer can verify the contents before publishing.