Reduce time complexity from O(n^2) to O(n) (#15417)
Improve topDownPath method performance changing the use of unshift for
push because from my investigation they have a time complexity of O(n)
and O(1) respectively in the different JavaScript engines. A recursive
approach would have a better readability in my opinion but speaking with
@CraigMacomber he told me that is better to keep it recursive in order
to avoid stack overflow errors.