refactor: Move generate readme command to separate project (#15396)
The `generate readme` command is a simple subclass of the oclif readme
command class with some minor tweaks. I want to move to ESM in the main
build-tools, but doing so is tricky with a command that subclasses an
existing class.

This change moves the readme command to a separate private package
within the release group, where it is used as a dev dependency within
the release group. This is safe because we don't need to distribute the
readme command; we only use it locally to generate the package readmes
within the build-tools release group.