refactor(client-debugger): Update library to use global singleton for managing the root devtools object (#15405)
Updates the devtools library to not require consumers to explicitly
dispose of their devtools object during app teardown. This has a couple
of major benefits:
1. Requiring consumers to dispose of the object puts substantial burden
on consumers to keep track of the devtools object in all relevant
places, which may or may not be reasonable for a (presumably)
dev-build-only entity.
2. We can ensure teardown occurs on page close / refresh, which helps
ensure the Chrome extension can clean up stale data.

This PR also updates the package's `packageDocumentation` comment and
README overview to reflect the new flow (and cleans up a significant
amount of outdated info).

See #15312 for an example of how this improves usage in application
flow.