Local checkpoints (#14627)
## Description

>In order to reduce strain on the group or "global" level MongoDB, this
change allows Scribe and Deli checkpoints to be written to the
local/operations DB while clients are actively working on a document.
When the last client disconnects, the checkpoint information for the
document is cleared from the local database and written to the global
database. This allows new sessions to start from this checkpoint.

> This pull request introduces a few changes to enable this
functionality:
> - Enables writing and reading from a new collection "checkpoints" in
the operations database.
> - When services restart, they restore checkpoints from either the
local or global database, depending on if clients were active.
> - These changes can be enabled/disabled with the flag
"localCheckpointEnabled"
> - It also includes some updates and improvements to logging in the
Scribe and Deli services.