fix(report): Handle workspace ranges when generating reports ()
Cherry-pick of #15439 to the build-tools release branch

---

The `release report` command was failing when run against a release
group using the workspace protocol. If the dependency is a workspace
dependency, then we need to use the current version of the package as
the dep range. Otherwise we pick the minimum version the range
represents.