[GC stress] Fail the test to validate new yaml pipeline (#15452)
Experimenting with changes to yaml pipeline by adding a job to create
work item on failure -
https://github.com/microsoft/FluidFramework/commit/ef9064a9b108e3b169f1ea2323ad216a86a229d5.

This change fails the test to see the new job works.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.