Fork treeTextFormat types (#15453)
## Description

Forks the format defined in `treeTextFormat.ts` into
`persistedTreeTextFormat.ts`. This allows hiding types which might end
up in the eventual persisted format from tree's public API. See
module-level doc comment on `persistedTreeTextFormat.ts` for more
details.

This change doesn't attempt to update references to the types in other
areas of the code, but it does add schemas for the persisted format.