feat: Support JSON output in check:changeset (#15465)
The changeset workflows continue to challenge me. The simple 0/1 output
isn't working for reasons I cannot explain, so I have added JSON support
to the check:changeset command. When the `--json` flag is passed, the
exit code will not be 1 in the case of a missing changeset. Instead the
JSON output includes a boolean value that should be checked.

I included the workflow changes in this PR so reviewers can see how
these command changes will be consumed. They won't be merged with this
PR - they'll be made separately (see #15472) after these changes are
available in the client release group.