Update the snapshot cache before closing (#15477)
[AB#4261](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4261)

Follow up to closing the summarizer client instead of refreshing it when
it knows it's in a bad state. We determined that the parent client that
was re-elected usually the same main client that spawned it.

The telemetry indicated summarizer was loading from the same cache on
average 4-5 times. Closing and loading new summarizers from the same
cache from the same main container has a negative performance impact on
the user.

This PR attempts to provide a solution to that problem by updating the
cache with the latest snapshot before closing the container.

Other options considered were explicitly telling the main client's
summary manager to load the next summarizer with the latest snapshot.
This was deemed harder to implement.