refactor(client-debugger-view): Rename package (#15470)
Renames the package to `@fluid-experimental/devtools-view`. Note the
change to the `fluid-experimental` namespace - this is intentional, in
preparation for publishing "alpha" releases. The package is still
private for now, and will be made public when we are ready to publish.

Also makes some corrections to the package README, though a larger audit
of that is needed.