Tree: Typed schema2 (#15459)
## Description

Redesigns the typed schema system to not rely on a central type
collection to generate schema aware types, as well as automatically
aggregate all types from a library using a builder.

This intentionally keeps the API similar to the old API to reduce
changes, and avoids adding much in the way of syntactic sugar to keep
things simple (it tends to complicate the type metaprogramming in ways
that interact poorly with recursive types and intelisense) and avoid
hiding concepts which are currently important to understand to use and
debug the system (like field kinds, primitives being objects etc).

## Breaking Changes

All use of the typed schema, or view schema in general is broken and
needs updating.

Renames:

1. FieldSchema --> FieldStoredSchema
2. TreeSchema --> TreeStoredSchema
3. jsonSchemaData --> jsonSchema (Also now a TypedSchemaColllection)
4. ValueFieldTreeFromSchema --> ValuePropertyFromSchema (since it is
unrelated to fields)
5. TypedSchema aggregator no longer exported. Many of its members moved
to InternalTypedSchemaTypes or exported unqualified. SchemaBuilder
replaces a lot of its use.