Implement SharedTree root local branch as a SharedTreeBranch (#15434)
## Description

The root local branch is currently managed by the EditManager. This PR
replaces it with an actual `SharedTreeBranch` instance, and updates the
relevant places to use operations on the branch instance rather than
having EditManager and SharedTreeCore re-implement them.

There are many **more** refactors that "fall out" of this initial
refactor. I am not attempting to do all of them here. In fact, just the
opposite - I am attempting to refrain from doing much more than the
minimal amount of changes that are required here as a first pass. Even
so, this PR has quite a few changes. I will follow this up with more
refactors in the future to continue to simplify our module boundaries
and reduce code.