feat(sequence): add fuzz test minimization (#15412)
Splits out fuzz test minimization code for sequence. It should be
possible to make this generic to other DDS's, but for now this code
focuses exclusively on sequence fuzz tests, as this reuses code from
before the generic DDS test harness existed.

Usage is similar to replaying specific seeds: change the test from
`describe.skip` to `describe.only` and update the seed variable.