offline: expose ability to get pending state without closing container (#15283)
Fix
[AB#4181](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4181)

Expose existing ability to stash pending local changes from
Container.closeAndGetPendingLocalState() but without closing the
container.

This increases the risk of stash blob misuse. In addition to not reusing
stash blobs, callers of this API must be careful not to use them after
the container has reconnected, i.e. when it emits a "connected" event.