Make undoRedoManager optional in branches (#15519)
## Description

In additional to various cleanups around the branch and UndoRedoManager
APIs, this PR makes undo/redo "optional" for branches. A branch may be
constructed without an UndoRedoManager, in which case it does not do any
of the undo/redo related tracking. This will allow future branches that
don't need undo/redo to be constructed without the overhead.