Only submit ops and signals when socket is connected (#15506)
## Description

Fix: https://dev.azure.com/fluidframework/internal/_workitems/edit/4174
Only submit ops and signals when socket is connected. This is to fix the
assert 0x244 which we were hitting in case connection is not yet
disposed but socket is already connected. This was happening in loop.
Scenario:
We have multiple containers in the given sessionId which share socket.
When socket disconnects all the containers need to go through
disconnection flow. During the disconnect flow of 1 container, when we
raise disconnect event on that container, the app tries to submit a
deletePresence signal on a different container which has not gone
through the disconnect flow yet and its connection is not yet disposed
while the socket is disconnected causing this assert and this exception
is recorded in the other container which emitted the "disconnect" event
which should be expected.

We should have dropped the signal in the happy path.