refactor(devtools-view): Refactor tree helper structure to reduce nested indentation (#15569)
FluentUI's experimental Tree component is very strange. It requires
setting up an outer `Tree` component context, which is really
(semantically) a "Tree Context". Inside of this context, an actual tree
is created using the `TreeItem` component (very unintuitive), and the
inline contents of that tree (the "header") are specified via the
`TreeItemLayout` component.

Because this is so counterintuitive, our current implementation
introduces a lot of ultimately unneeded component hierarchy, which
introduces a lot of unwanted and unpredictable indentation.

These changes are aimed at cleaning up that indentation, which it
accomplishes by re-structuring our app tree. Since the setup is
unintuitive, I went ahead and made our tree helpers an implementation
detail of our data-visualization components directory. If we want to try
and generalize this later, we can, but for now I am opting out of
sharing.

This PR also:
* Encapsulates `node` prop accepted by the various tree visualization
components into a shared interface.
* Adds `@fluentui/react-components/unstable` to an allow-list in the
eslint config so we don't have to disable line-by-line.
* Removes usages of `nodeKind` in user-facing visualization (this data
is an implementation detail of the visualization tree format).
* Adds labels to unknown data type tree nodes


![image](https://github.com/microsoft/FluidFramework/assets/54606601/b0189d23-5547-45a9-ada3-c02e9d3bdfff)