Remove sessionId from GraphCommit (#15578)
## Description

`sessionId` is a relatively redundant field on `GraphCommit`s. With the
exception of commits on the trunk, adjacent commits in branches all have
the same sessionId. Removing this field from `SharedTreeBranch` has the
additional benefit of making it possible to model the trunk as a
`SharedTreeBranch` in the future.

This does not change the persisted data format of the summarized
`EditManager`, athough the data format can now be simplified in the
future by not including the session ID in the peer branches.