Update EditManager trunk to use SharedTreeBranch (#15583)
## Description

This replaces `EditManager`s head trunk commit with a
`SharedTreeBranch`. The branch methods are not fully leveraged yet (for
example, `pushToTrunk` resorts to a call to `setHead`), but we may be
able to improve this after a future refactor of `UndoRedoManager`. Since
`SharedTreeBranches` do not track session IDs or sequence numbers,
`TrunkCommit` has been removed in favor of a trunk metadata lookup map
that serves the same purpose. There is also some minor cleanup to a test
method and some removal of whitespace in private docs.