[GC] Refactor Garbage collector by moving telemetry tracking to a helper class (#15535)
## Description
Moved our the telemetry specific logic in Garbage Collector to a
separate helper class. GC has a complicated logic to identify and log
when an object is accessed after it might have been GC'd. It does this
as a best case effort and the logic is separate for summarizer and
non-summarizer clients. As such, it is not really a part of the core GC
functionality but which is needed for now while there are GC bugs. So,
moving it out to a separate class has the following advantages:
- It makes unit testing the telemetry logic better / easier. I added a
bunch of unit tests, some of which already existed (in
garbageCollection.spec.ts) and others are new (like
gcUnknownOutboundReferences tests). This makes adding more test coverage
easier.
- The main Garbage Collector class deals with the core part of the GC
logic, making it easier to follow and understand. The same goes for the
telemetry part which is now encapsulated on its own.

## Reviewer Guidance
This is a refactor. There is no change in functionality.