build(devtools-browser-extension): Remove unnecessary file copy step (#15600)
The TSC build is only used to enable Mocha testing in the package, so
the build script that copied json, html, and css assets is not
necessary. Additionally, it was copying the `fence.json` files we use to
enforce directory-wise dependency invariants into the build output,
which was not intended.