Remove deprecated items from container-runtime-definitions and aqueduct (#15570)
This change removes three deprecated items:

1. "localHelp" event on IContainerRuntimeEvents (#15518 followup)
2. PureDataObject.getFluidObjectFromDirectory (#15530 followup)
3. IProvideContainerRuntime (#8299 followup)

For the last one, I also remove the automatic registration of the
container runtime in the BaseContainerRuntimeFactory's dependency
container (this seems inappropriate to me anyway) and removed the
property from `ContainerRuntime` as well. I've proactively removed
usages of the property from the Loop app.